requirements.txt
torch
torchaudio
pysoundfile
huggingface_hub
safetensors
gradio
numpy